+++ src/dto/index.ts
@@ -1,0 +1,5 @@
+export * from "./SubjectDto";
+
+export * from "./TaxonomyDto";
+
+export * from "./WorkDto";