+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from "./core";
 
-export * from "./json";
+export * from "./dto";
 
 export * from "./raw";