+++ src/json/index.ts
@@ -1,5 +1,0 @@
-export * from "./SubjectJson";
-
-export * from "./TaxonomyJson";
-
-export * from "./WorkJson";